+++ src/modules/GridControl.hpp
@@ -11,8 +11,6 @@ class $editorModule(GridControl) {
 
     void onEditor();
     void updateGrid(float newValue = 0, bool updateInput = true);
-    static float getSettingsMenuWidth(EditorUI* editorUI);
-    static float getSliderMaxX(EditorUI* editorUI);
     static float getSliderMinY(EditorUI* editorUI);
 
     GridControl();